Add files via upload
This are some test photos of the base design. Take a look. Finishing up the test cases we might run into. This is a rough draft. Maybe this week or next week we can go over what you guys might want instead. I will start by making log in page and then we can use that as a base for the rest. Placing this in a branch so if not the final draft we can edit it.